upgrade to SDK 0.13.20 Include logback.xml to avoid log DEBUG spamming Update README : dar file output from daml new project has slightly different name avoid bundle ledger-api-test-tool.jar with the project Rename examples to damlonfabric examples
