+++ build.sbt
@@ -5,7 +5,7 @@ ThisBuild / version          := "1.1"
 ThisBuild / organization     := "com.hacera"
 ThisBuild / organizationName := "HACERA"
 
-lazy val sdkVersion = "100.13.12"
+lazy val sdkVersion = "100.13.20"
 
 // This task is used by the integration test to detect which version of Ledger API Test Tool to use.
 val printSdkVersion= taskKey[Unit]("printSdkVersion")
@@ -33,11 +33,14 @@ lazy val root = (project in file("."))
       "com.typesafe.akka" %% "akka-stream" % "2.5.22",
       "com.typesafe.akka" %% "akka-stream-testkit" % "2.5.22" % Test,
       "com.typesafe.akka" %% "akka-slf4j" % "2.5.22",
-      "org.slf4j" % "slf4j-api" % "1.7.25" % "provided",
-      "org.slf4j" % "slf4j-simple" % "1.7.25" % "provided",
+      "org.slf4j" % "slf4j-api" % "1.7.26",
+      "ch.qos.logback" % "logback-core" % "1.2.3",
+      "ch.qos.logback" % "logback-classic" % "1.2.3",
+      "commons-io" % "commons-io" % "2.6",
+      "com.digitalasset" % "daml-lf-archive" % sdkVersion,
       "com.digitalasset.ledger" %% "ledger-api-common" % sdkVersion,
       "com.daml.ledger" %% "api-server-damlonx" % sdkVersion,
-      "com.daml.ledger" %% "participant-state-index" % sdkVersion,
+      "com.daml.ledger" %% "participant-state-index-v1" % sdkVersion,
       "com.daml.ledger" %% "reference-participant-state-index" % sdkVersion,
       "com.daml.ledger" %% "participant-state" % sdkVersion,
       "com.daml.ledger" %% "participant-state-kvutils" % sdkVersion,
