+++ src/main/scala/com/hacera/FabricParticipantState.scala
@@ -20,10 +20,8 @@ import com.digitalasset.daml.lf.engine.Engine
 import com.digitalasset.daml_lf.DamlLf.Archive
 import com.digitalasset.platform.akkastreams.dispatcher.Dispatcher
 import com.digitalasset.platform.akkastreams.dispatcher.SubSource.OneAfterAnother
-import com.daml.ledger.participant.state.kvutils.KeyValueSubmission
+import com.daml.ledger.participant.state.kvutils.{KeyValueCommitting, KeyValueConsumption, KeyValueSubmission, Pretty}
 import com.daml.ledger.participant.state.backport.TimeModel
-import com.daml.ledger.participant.state.kvutils.KeyValueCommitting
-import com.daml.ledger.participant.state.kvutils.KeyValueConsumption
 import com.google.protobuf.ByteString
 import org.slf4j.LoggerFactory
 
@@ -71,7 +69,10 @@ class FabricParticipantState(roleTime: Boolean, roleLedger: Boolean, participant
 
   // The ledger configuration
   private val ledgerConfig = Configuration(
-    TimeModel(JDuration.ofSeconds(600L), JDuration.ofSeconds(600L), JDuration.ofSeconds(600L)).get
+    1L,
+    TimeModel(JDuration.ofSeconds(600L), JDuration.ofSeconds(600L), JDuration.ofSeconds(600L)).get,
+    authorizedParticipantId = Some(participantId),
+    openWorld = true
   )
 
   // DAML Engine for transaction validation.
@@ -103,14 +104,18 @@ class FabricParticipantState(roleTime: Boolean, roleLedger: Boolean, participant
   /** Reference to the latest state.
     */
   @volatile private var stateRef: State =
-    State(
-      //commitLog = Vector.empty[Commit],
-      //recordTime = Timestamp.Epoch,
-      //store = Map.empty[ByteString, ByteString],
-      config = Configuration(
-        timeModel = TimeModel.reasonableDefault
-      )
+  State(
+    //commitLog = Vector.empty[Commit],
+    //recordTime = Timestamp.Epoch,
+    //store = Map.empty[ByteString, ByteString],
+    //timeModel = TimeModel.reasonableDefault
+    config = Configuration(
+      1L,
+      timeModel = TimeModel.reasonableDefault,
+      authorizedParticipantId = Some(participantId),
+      openWorld = true
     )
+  )
 
   private def serializeCommit(commit: Commit): Array[Byte] = {
 
@@ -166,18 +171,17 @@ class FabricParticipantState(roleTime: Boolean, roleLedger: Boolean, participant
           logger.warn(s"CommitActor: duplicate entry identifier in commit message, ignoring.")
         } else {
           logger.trace(
-            s"CommitActor: processing submission ${KeyValueCommitting.prettyEntryId(entryId)}..."
+            s"CommitActor: processing submission ${Pretty.prettyEntryId(entryId)}..."
           )
           // Process the submission to produce the log entry and the state updates.
           //this.synchronized {
           val (logEntry, damlStateUpdates) = KeyValueCommitting.processSubmission(
             engine,
-            state.config,
             entryId,
             newRecordTime,
+            state.config,
             submission,
-            submission.getInputLogEntriesList.asScala
-              .map(eid => eid -> getLogEntry(state, eid))(breakOut),
+            participantId,
             submission.getInputDamlStateList.asScala
               .map(key => key -> getDamlState(state, key))(breakOut)
           )
@@ -191,7 +195,7 @@ class FabricParticipantState(roleTime: Boolean, roleLedger: Boolean, participant
             } + (entryId.getEntryId -> KeyValueCommitting.packDamlLogEntry(logEntry))
 
           logger.trace(
-            s"CommitActor: committing ${KeyValueCommitting.prettyEntryId(entryId)} and ${allUpdates.size} updates to store."
+            s"CommitActor: committing ${Pretty.prettyEntryId(entryId)} and ${allUpdates.size} updates to store."
           )
 
           // Write some state to Fabric
@@ -504,10 +508,20 @@ class FabricParticipantState(roleTime: Boolean, roleLedger: Boolean, participant
 
   /** Allocate a party on the ledger */
   override def allocateParty(
-      hint: Option[String],
-      displayName: Option[String]
-  ): CompletionStage[PartyAllocationResult] =
-    // TODO: Implement party management (does not work yet just like in reference)
+                              hint: Option[String],
+                              displayName: Option[String]
+                            ): CompletionStage[PartyAllocationResult] =
+  // TODO: Implement party management (does not work yet just like in reference)
     CompletableFuture.completedFuture(PartyAllocationResult.NotSupported)
 
+  override def submitConfiguration(maxRecordTime: Timestamp, submissionId: String, config: Configuration): CompletionStage[SubmissionResult] =
+    CompletableFuture.completedFuture({
+      val submission =
+        KeyValueSubmission.configurationToSubmission(maxRecordTime, submissionId, config)
+      commitActorRef ! CommitSubmission(
+        allocateEntryId,
+        submission
+      )
+      SubmissionResult.Acknowledged
+    })
 }
